edge/data-preprocessor/scripts/requirements.txt
setuptools
paho-mqtt==1.2.3
pika